requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2
django-admin-honeypot==1.1.0
django-braces==1.14.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-log-viewer==1.1.2
django-oauth-toolkit==1.5.0
django-oauth2-provider==0.2.6.1
django-oidc-provider==0.7.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
djangorestframework-oauth==1.1.0
drf-yasg==1.20.0
future==0.18.2
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==7.0.0

psycopg2==2.8.6
pycparser==2.20
pycryptodomex==3.10.1
pyjwkest==1.4.2
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
shortuuid==1.0.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.12.1
